// projects/library/rbac-service/src/lib/tsc/enum/rbac.enum.ts
export enum ApplicationResourcePermission {
    NETZERO_ADMIN_RESOURCE_PERMISSION = 'NETZERO_ADMIN_RESOURCE_PERMISSION',
    ADMIN_RESOURCE_PERMISSION = 'ADMIN_RESOURCE_PERMISSION',
    NETZERO_RESOURCE_PERMISSION = 'NETZERO_RESOURCE_PERMISSION',
    ENERGY_RESOURCE_PERMISSION = 'ENERGY_RESOURCE_PERMISSION',
    WATER_RESOURCE_PERMISSION = 'WATER_RESOURCE_PERMISSION',
    ENVIRO_RESOURCE_PERMISSION = 'ENVIRO_RESOURCE_PERMISSION',
    CONNECT_RESOURCE_PERMISSION = 'CONNECT_RESOURCE_PERMISSION',
    OFFSET_RESOURCE_PERMISSION = 'OFFSET_RESOURCE_PERMISSION',
    SUPPLIER_RESOURCE_PERMISSION = 'SUPPLIER_RESOURCE_PERMISSION',
    EMPLOYEE_RESOURCE_PERMISSION = 'EMPLOYEE_RESOURCE_PERMISSION'
}